requirements.txt
SQLAlchemy~=1.4.22
fastapi~=0.67.0
python-multipart~=0.0.5
aiofiles~=0.7.0
celery~=5.1.2
alembic~=1.6.5
uvicorn~=0.14.0
redis~=3.5.3
asyncpg~=0.23.0
psycopg2~=2.9.1
python-dotenv~=0.19.0